requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
firebase-admin==2.16.0
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
google-api-core==1.10.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud-core==0.29.1
google-cloud-firestore==1.1.0
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
grpcio==1.20.1
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
oauth2client==4.1.3
oauthlib==3.0.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.3
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
